Allow user to increase font size
Add a settings section, which for now only supports font size. But can support other things as needed, stored in the settings object. The default font is quite small IMO, so I think making it a setting will be nice.

I also removed the hard-coded body height, which allows for the bottom padding of `body` to be applied.